Fix hashmap spill mechanism for removing/finding values by their key
Will probably implement buckets later on instead of using this spill over mechanism. You can probably keep the spill mechanism in check with a good hashing function and carefully selected heuristics about when to resize the hashmap, but buckets of linked lists seem to be better imo.
Now passing in the key to be hashed for the hashmap properly by getting the pointer to the key value (the hashing function dereferences it again).
Also fixed the releasing of tracked resources.